feat: add CookieUtil for JWT cookie generation and update gitignore for local SonarQube script

+++ src/main/java/org/me/dibs/config/CookieUtil.java
@@ -11,7 +11,7 @@ public class CookieUtil {
                 .secure(true)
                 .sameSite("None")
                 .path("/")
-                .maxAge(24 * 60 * 60)
+                .maxAge((long)24 * 60 * 60)
                 .build();
     }
 }
